+++ examples/equipment_unit/_api_client.py
@@ -24,7 +24,7 @@ class EquipmentUnitClient:
 
     Generated with:
         pygen = 0.99.35
-        cognite-sdk = 7.62.5
+        cognite-sdk = 7.62.6
         pydantic = 2.9.2
 
     Data Model: