+++ examples/omni/_api_client.py
@@ -42,7 +42,7 @@ class OmniClient:
 
     Generated with:
         pygen = 0.99.35
-        cognite-sdk = 7.62.5
+        cognite-sdk = 7.62.6
         pydantic = 2.9.2
 
     Data Model: