+++ examples/scenario_instance/client/_api_client.py
@@ -22,7 +22,7 @@ class ScenarioInstanceClient:
 
     Generated with:
         pygen = 0.99.35
-        cognite-sdk = 7.62.5
+        cognite-sdk = 7.62.6
         pydantic = 2.9.2
 
     Data Model: