+++ docs/CHANGELOG.md
@@ -13,6 +13,10 @@ Changes are grouped as follows
 - `Fixed` for any bug fixes.
 - `Security` in case of vulnerabilities.
 
+## TBD
+### Removed
+- `pydantic` `v1` is no longer supported. This is because `pyodide` now supports `pydantic` `2.7`.
+
 ## [0.99.35] - 24-09-30
 ### Added
 - Support for `Enum` in `MockGenerator`.